fixed 'new' handler, working on saving of nb

+++ IPython/frontend/html/notebook/filenbmanager.py
@@ -147,10 +147,13 @@ class FileNotebookManager(NotebookManager):
         #name = self.get_name(notebook_id)
         return self.get_path_by_name(notebook_name, notebook_path)
 
-    def get_path_by_name(self, name, notebook_path):
+    def get_path_by_name(self, name, notebook_path=None):
         """Return a full path to a notebook given its name."""
         filename = name #+ self.filename_ext
-        path = os.path.join(self.notebook_dir, notebook_path, filename)
+        if notebook_path == None:
+            path = os.path.join(self.notebook_dir, filename)
+        else:    
+            path = os.path.join(self.notebook_dir, notebook_path, filename)
         return path
 
     def read_notebook_object_from_path(self, path):
@@ -169,7 +172,6 @@ class FileNotebookManager(NotebookManager):
     def read_notebook_object(self, notebook_name, notebook_path):
         """Get the Notebook representation of a notebook by notebook_id."""
         path = self.get_path(notebook_name, notebook_path)
-        self.log.info(path)
         if not os.path.isfile(path):
             raise web.HTTPError(404, u'Notebook does not exist: %s' % notebook_name)
         last_modified, nb = self.read_notebook_object_from_path(path)
@@ -177,23 +179,19 @@ class FileNotebookManager(NotebookManager):
         nb.metadata.name = os.path.splitext(os.path.basename(path))[0]
         return last_modified, nb
     
-    def write_notebook_object(self, nb, notebook_id=None):
-        """Save an existing notebook object by notebook_id."""
+    def write_notebook_object(self, nb, notebook_name=None, notebook_path=None):
+        """Save an existing notebook object by notebook_name."""
         try:
             new_name = nb.metadata.name
         except AttributeError:
             raise web.HTTPError(400, u'Missing notebook name')
-
-        if notebook_id is None:
-            notebook_id = self.new_notebook_id(new_name)
-
-        if notebook_id not in self.mapping:
-            raise web.HTTPError(404, u'Notebook does not exist: %s' % notebook_id)
-
-        old_name = self.mapping[notebook_id]
-        old_checkpoints = self.list_checkpoints(notebook_id)
         
-        path = self.get_path_by_name(new_name)
+        new_path = notebook_path
+        old_name = notebook_name
+#        old_name = self.mapping[notebook_name]
+#        old_checkpoints = self.list_checkpoints(notebook_id)
+        
+        path = self.get_path_by_name(new_name, new_path)
         try:
             self.log.debug("Autosaving notebook %s", path)
             with open(path,'w') as f:
@@ -211,36 +209,34 @@ class FileNotebookManager(NotebookManager):
             except Exception as e:
                 raise web.HTTPError(400, u'Unexpected error while saving notebook as script: %s' % e)
         
-        # remove old files if the name changed
-        if old_name != new_name:
-            # update mapping
-            self.mapping[notebook_id] = new_name
-            self.rev_mapping[new_name] = notebook_id
-            del self.rev_mapping[old_name]
+        if old_name != None:
+            # remove old files if the name changed
+            if old_name != new_name:            
+                # remove renamed original, if it exists
+                old_path = self.get_path_by_name(old_name, notebook_path)
+                if os.path.isfile(old_path):
+                    self.log.debug("unlinking notebook %s", old_path)
+                    os.unlink(old_path)
             
-            # remove renamed original, if it exists
-            old_path = self.get_path_by_name(old_name)
-            if os.path.isfile(old_path):
-                self.log.debug("unlinking notebook %s", old_path)
-                os.unlink(old_path)
+                # cleanup old script, if it exists
+                if self.save_script:
+                    old_pypath = os.path.splitext(old_path)[0] + '.py'
+                    if os.path.isfile(old_pypath):
+                        self.log.debug("unlinking script %s", old_pypath)
+                        os.unlink(old_pypath)
+            """    
+                # rename checkpoints to follow file
+                for cp in old_checkpoints:
+                    checkpoint_id = cp['checkpoint_id']
+                    old_cp_path = self.get_checkpoint_path_by_name(old_name, checkpoint_id)
+                    new_cp_path = self.get_checkpoint_path_by_name(new_name, checkpoint_id)
+                    if os.path.isfile(old_cp_path):
+                        self.log.debug("renaming checkpoint %s -> %s", old_cp_path, new_cp_path)
+                        os.rename(old_cp_path, new_cp_path)
+            """
+        return new_name
             
-            # cleanup old script, if it exists
-            if self.save_script:
-                old_pypath = os.path.splitext(old_path)[0] + '.py'
-                if os.path.isfile(old_pypath):
-                    self.log.debug("unlinking script %s", old_pypath)
-                    os.unlink(old_pypath)
-            
-            # rename checkpoints to follow file
-            for cp in old_checkpoints:
-                checkpoint_id = cp['checkpoint_id']
-                old_cp_path = self.get_checkpoint_path_by_name(old_name, checkpoint_id)
-                new_cp_path = self.get_checkpoint_path_by_name(new_name, checkpoint_id)
-                if os.path.isfile(old_cp_path):
-                    self.log.debug("renaming checkpoint %s -> %s", old_cp_path, new_cp_path)
-                    os.rename(old_cp_path, new_cp_path)
-            
-        return notebook_id
+
 
     def delete_notebook(self, notebook_id):
         """Delete notebook by notebook_id."""
@@ -261,7 +257,7 @@ class FileNotebookManager(NotebookManager):
         os.unlink(nb_path)
         self.delete_notebook_id(notebook_id)
 
-    def increment_filename(self, basename):
+    def increment_filename(self, basename, notebook_path=None):
         """Return a non-used filename of the form basename<int>.
         
         This searches through the filenames (basename0, basename1, ...)
@@ -270,8 +266,8 @@ class FileNotebookManager(NotebookManager):
         """
         i = 0
         while True:
-            name = u'%s%i' % (basename,i)
-            path = self.get_path_by_name(name)
+            name = u'%s%i.ipynb' % (basename,i)
+            path = self.get_path_by_name(name, notebook_path)
             if not os.path.isfile(path):
                 break
             else:
@@ -280,28 +276,28 @@ class FileNotebookManager(NotebookManager):
     
     # Checkpoint-related utilities
     
-    def get_checkpoint_path_by_name(self, name, checkpoint_id):
+    def get_checkpoint_path_by_name(self, name, checkpoint_name):
         """Return a full path to a notebook checkpoint, given its name and checkpoint id."""
-        filename = "{name}-{checkpoint_id}{ext}".format(
+        filename = "{name}-{checkpoint_name}{ext}".format(
             name=name,
-            checkpoint_id=checkpoint_id,
+            checkpoint_name=checkpoint_name,
             ext=self.filename_ext,
         )
         path = os.path.join(self.checkpoint_dir, filename)
         return path
     
-    def get_checkpoint_path(self, notebook_id, checkpoint_id):
+    def get_checkpoint_path(self, notebook_name, checkpoint_name):
         """find the path to a checkpoint"""
-        name = self.get_name(notebook_id)
-        return self.get_checkpoint_path_by_name(name, checkpoint_id)
+        name = notebook_name
+        return self.get_checkpoint_path_by_name(name, checkpoint_name)
     
-    def get_checkpoint_info(self, notebook_id, checkpoint_id):
+    def get_checkpoint_info(self, notebook_name, checkpoint_name):
         """construct the info dict for a given checkpoint"""
-        path = self.get_checkpoint_path(notebook_id, checkpoint_id)
+        path = self.get_checkpoint_path(notebook_name, checkpoint_name)
         stats = os.stat(path)
         last_modified = datetime.datetime.utcfromtimestamp(stats.st_mtime)
         info = dict(
-            checkpoint_id = checkpoint_id,
+            checkpoint_name = checkpoint_name,
             last_modified = last_modified,
         )
         
@@ -309,54 +305,54 @@ class FileNotebookManager(NotebookManager):
         
     # public checkpoint API
     
-    def create_checkpoint(self, notebook_id):
+    def create_checkpoint(self, notebook_name):
         """Create a checkpoint from the current state of a notebook"""
-        nb_path = self.get_path(notebook_id)
+        nb_path = self.get_path(notebook_name)
         # only the one checkpoint ID:
-        checkpoint_id = "checkpoint"
-        cp_path = self.get_checkpoint_path(notebook_id, checkpoint_id)
-        self.log.debug("creating checkpoint for notebook %s", notebook_id)
+        checkpoint_name = "checkpoint"
+        cp_path = self.get_checkpoint_path(notebook_name, checkpoint_name)
+        self.log.debug("creating checkpoint for notebook %s", notebook_name)
         if not os.path.exists(self.checkpoint_dir):
             os.mkdir(self.checkpoint_dir)
         shutil.copy2(nb_path, cp_path)
         
         # return the checkpoint info
-        return self.get_checkpoint_info(notebook_id, checkpoint_id)
+        return self.get_checkpoint_info(notebook_name, checkpoint_name)
     
-    def list_checkpoints(self, notebook_id):
+    def list_checkpoints(self, notebook_name):
         """list the checkpoints for a given notebook
         
         This notebook manager currently only supports one checkpoint per notebook.
         """
-        checkpoint_id = "checkpoint"
-        path = self.get_checkpoint_path(notebook_id, checkpoint_id)
+        checkpoint_name = "checkpoint"
+        path = self.get_checkpoint_path(notebook_name, checkpoint_name)
         if not os.path.exists(path):
             return []
         else:
-            return [self.get_checkpoint_info(notebook_id, checkpoint_id)]
+            return [self.get_checkpoint_info(notebook_name, checkpoint_name)]
         
     
-    def restore_checkpoint(self, notebook_id, checkpoint_id):
+    def restore_checkpoint(self, notebook_name, checkpoint_name):
         """restore a notebook to a checkpointed state"""
-        self.log.info("restoring Notebook %s from checkpoint %s", notebook_id, checkpoint_id)
-        nb_path = self.get_path(notebook_id)
-        cp_path = self.get_checkpoint_path(notebook_id, checkpoint_id)
+        self.log.info("restoring Notebook %s from checkpoint %s", notebook_name, checkpoint_name)
+        nb_path = self.get_path(notebook_name)
+        cp_path = self.get_checkpoint_path(notebook_name, checkpoint_name)
         if not os.path.isfile(cp_path):
             self.log.debug("checkpoint file does not exist: %s", cp_path)
             raise web.HTTPError(404,
-                u'Notebook checkpoint does not exist: %s-%s' % (notebook_id, checkpoint_id)
+                u'Notebook checkpoint does not exist: %s-%s' % (notebook_name, checkpoint_name)
             )
         # ensure notebook is readable (never restore from an unreadable notebook)
         last_modified, nb = self.read_notebook_object_from_path(cp_path)
         shutil.copy2(cp_path, nb_path)
         self.log.debug("copying %s -> %s", cp_path, nb_path)
     
-    def delete_checkpoint(self, notebook_id, checkpoint_id):
+    def delete_checkpoint(self, notebook_name, checkpoint_name):
         """delete a notebook's checkpoint"""
-        path = self.get_checkpoint_path(notebook_id, checkpoint_id)
+        path = self.get_checkpoint_path(notebook_name, checkpoint_name)
         if not os.path.isfile(path):
             raise web.HTTPError(404,
-                u'Notebook checkpoint does not exist: %s-%s' % (notebook_id, checkpoint_id)
+                u'Notebook checkpoint does not exist: %s-%s' % (notebook_name, checkpoint_name)
             )
         self.log.debug("unlinking %s", path)
         os.unlink(path)
